+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/docs/As12PCodeDocs.java
@@ -164,71 +164,7 @@ public class As12PCodeDocs extends AbstractDocs {
             }
         }
         return identName.toString();
-    }
-
-    private static String hilightArgument(String docs, int argumentIndex) {
-        if (argumentIndex < 0) {
-            return docs;
-        }
-        String opHeader = "<span class=\"instruction-operands\">";
-        int opIndex = docs.indexOf(opHeader) + opHeader.length();
-        int opEndIndex = docs.indexOf("</span>", opIndex);
-        String operandDocs = docs.substring(opIndex, opEndIndex).trim();
-        if (operandDocs.isEmpty()) {
-            return docs;
-        }
-
-        DocsOperandLexer lexer = new DocsOperandLexer(new StringReader(operandDocs));
-        try {
-            ParsedSymbol symb;
-
-            int pos = 0;
-            int endPos = 0;
-            int startPos = 0;
-            while (true) {
-                startPos = lexer.yychar();
-                symb = lexer.lex();
-                if (symb.type == ParsedSymbol.TYPE_BRACKET_OPEN) {
-                    while (symb.type != ParsedSymbol.TYPE_BRACKET_CLOSE && symb.type != ParsedSymbol.TYPE_EOF) {
-                        symb = lexer.lex();                        
-                    }
-                    endPos = lexer.yychar() + 1;
-                    break;
-                }
-                if (symb.type == ParsedSymbol.TYPE_IDENTIFIER) {
-                    symb = lexer.lex();
-                    endPos = lexer.yychar();
-                    if (symb.type == ParsedSymbol.TYPE_COLON) {
-                        do {
-                            symb = lexer.lex();
-                            if (symb.type != ParsedSymbol.TYPE_IDENTIFIER && symb.type != ParsedSymbol.TYPE_STAR) {
-                                throw new IOException("type identifier expected");
-                            }
-                            symb = lexer.lex();
-                            endPos = lexer.yychar();
-                        } while (symb.type == ParsedSymbol.TYPE_PIPE);
-                    }
-
-                    if (pos == argumentIndex) {
-                        break;
-                    }
-
-                    if (symb.type == ParsedSymbol.TYPE_COMMA) {
-                        pos++;
-                    } else {
-                        break;
-                    }
-                }
-            }
-            String hilightedOperandDocs = operandDocs.substring(0, startPos)
-                    + "<strong class=\"selected-operand\">" + operandDocs.substring(startPos, endPos) + "</strong>"
-                    + operandDocs.substring(endPos);
-            docs = docs.substring(0, opIndex) + hilightedOperandDocs + docs.substring(opEndIndex);
-        } catch (IOException ex) {
-            //ignore
-        }
-        return docs;
-    }
+    }   
 
     public static String getDocsForIns(String insName, boolean ui, boolean standalone, boolean nightMode, int argumentToHilight) {
         insName = insName.toLowerCase();